+++ gazebo/transport/Connection.cc
@@ -86,7 +86,6 @@ Connection::Connection()
 {
   this->isOpen = false;
   this->dropMsgLogged = false;
-  this->headerBuffer = new char[HEADER_LENGTH+1];
 
   if (iomanager == NULL)
     iomanager = new IOManager();
@@ -122,9 +121,6 @@ Connection::Connection()
 //////////////////////////////////////////////////
 Connection::~Connection()
 {
-  delete [] this->headerBuffer;
-  this->headerBuffer = NULL;
-
   this->Shutdown();
 
   if (iomanager)
@@ -303,7 +299,8 @@ void Connection::EnqueueMsg(const std::string &_buffer,
     return;
   }
 
-  snprintf(this->headerBuffer, HEADER_LENGTH + 1, "%08x",
+  char headerBufferLocal[HEADER_LENGTH + 1];
+  snprintf(headerBufferLocal, HEADER_LENGTH + 1, "%08x",
       static_cast<unsigned int>(_buffer.size()));
 
   {
@@ -312,9 +309,9 @@ void Connection::EnqueueMsg(const std::string &_buffer,
     if (this->writeQueue.empty() ||
         (this->writeCount > 0 && this->writeQueue.size() == 1) ||
         (this->writeQueue.back().size()+_buffer.size() > 4096))
-      this->writeQueue.push_back(std::string(headerBuffer) + _buffer);
+      this->writeQueue.push_back(std::string(headerBufferLocal) + _buffer);
     else
-      this->writeQueue.back() += std::string(headerBuffer) + _buffer;
+      this->writeQueue.back() += std::string(headerBufferLocal) + _buffer;
     this->callbacks.push_back(std::make_pair(_cb, _id));
   }